% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/something.R
\name{marcela_func}
\alias{marcela_func}
\title{This function is awesome!}
\usage{
marcela_func(x)
}
\arguments{
\item{x}{a number}
}
\description{
This function is awesome!
}
\details{
This function does awesome things. It's too complex for you to understand.
}
